Split test/test_components.py into test/components/test_*.py
