+++ test/components/conftest.py
@@ -1,0 +1,3 @@
+import os
+
+os.environ["GRADIO_ANALYTICS_ENABLED"] = "False"
